refactor content parsing
